Stricter Metabuli parsers
The Metabuli outputs are obstensibly in TSV format, but has been changed in
Metabuli commit a17debb, and may have been changed in previous versions.
Therefore, we a) need to handle both versions of the files, depending on what
version of Metabuli is used, and b) need to guard against future changes to the
format.

This commit rewrites the parser to be stricter.

It also deletes a seemingly unused zero-arg function `metabuli_lineage()`, and
the unused constant METABULI_LINEAGE.